feature: main-service 1-iteration

+++ main-service/src/main/java/ru/practicum/service/CategoryServiceImpl.java
@@ -5,10 +5,13 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.CategoryDto;
+import ru.practicum.handler.exception.ConflictException;
 import ru.practicum.handler.exception.ObjectNotFoundException;
 import ru.practicum.mapper.CategoryMapper;
 import ru.practicum.model.Category;
+import ru.practicum.model.Event;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -17,12 +20,18 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Transactional ///
     @Override
     public CategoryDto saveCategory(CategoryDto categoryDto) {
-        Category category = categoryRepository.save(CategoryMapper.toDtoCategory(categoryDto));
-        return CategoryMapper.toCategoryDto(category);
+        String name = categoryDto.getName();
+        if (categoryRepository.findByName(name).isEmpty()) {
+            Category category = categoryRepository.save(CategoryMapper.toDtoCategory(categoryDto));
+            return CategoryMapper.toCategoryDto(category);
+        } else {
+            throw new ConflictException("Категория с именем " + name + " уже существует.");
+        }
     }
 
     @Transactional ///
@@ -30,8 +39,14 @@ public class CategoryServiceImpl implements CategoryService {
     public CategoryDto updateCategory(CategoryDto categoryDto, Long catId) {
         Category category = categoryRepository.findById(catId).orElseThrow(() -> new ObjectNotFoundException("Category not found"));
 
-        if (categoryDto.getName() != null) {
-            category.setName(categoryDto.getName());
+        String name = categoryDto.getName();
+        if (name != null) {
+            Category categoryByName = categoryRepository.findByName(name).orElse(null);
+            if (categoryByName == null || categoryByName.getId().equals(catId)) {
+                category.setName(categoryDto.getName());
+            } else {
+                throw new ConflictException("Категория с именем " + name + " уже существует.");
+            }
         }
         return CategoryMapper.toCategoryDto(categoryRepository.save(category));
     }
@@ -52,6 +67,12 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional ///
     @Override
     public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+        Category category = categoryRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Category not found"));
+        List<Event> events = eventRepository.findBycategory(category);
+        if (events.size() == 0) {
+            categoryRepository.deleteById(id);
+        } else {
+            throw new ConflictException("Нельзя удалить не пустую категорию.");
+        }
     }
 }
